fix 2 pass workflow
purpose of 2 pass is to scan build ids from top to bottom then vice-versa.
this allows us to apply last known totals to jobs that have been aborted
and were not able to report total number of tests.

the fix here is to pass in lastTotalCount into 2nd pass so that it can be
used if detected after bottom-up scan.  also this patch will delete
stale versions of a job so that only the newest run against a version is used