CBQE-6225: Incorrect failure count reported for Aborted and failed tests
This changes the behaviour when no totalCount is reported which can happen if the job is aborted or fails. Currently, the totalCount is set to the count of the last run if it exists. In the case of an executor job, the last run is likely to have been with completely different parameters. This results in a different totalCount being set each time a job fails based on whatever jobs also finished in the same poll interval. With this change, executor jobs with 0 totalCount are stored as 0 totalCount and non executor jobs with 0 totalCount are stored as the totalCount of the last run if it exists, otherwise it is skipped.

Change-Id: I1da338a9213576f4110944392176dd9a646940e3
Reviewed-on: http://review.couchbase.org/c/jinja/+/142031
Reviewed-by: Jagadesh Munta <jagadesh.munta@couchbase.com>
Tested-by: Jake Rawsthorne <jake.rawsthorne@couchbase.com>